django/groupstudy_dashboard/uploads/views.py
from rest_framework import viewsets
from . import models, serializers


class DocumentViewSet(viewsets.ModelViewSet):
    queryset = models.Document.objects.all()
    serializer_class = serializers.DocumentSerializer
