django/requirements.txt
django==1.10
djangorestframework=3.5.2
